% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_exps}
\alias{combine_exps}
\title{Combine and compare proteomics experiments.}
\usage{
combine_exps(exp1, exp2, exp1_id, exp2_id, outfile_prefix)
}
\arguments{
\item{exp1}{A data frame produced by \code{combine_reps()} corresponding to experiment 1.}

\item{exp2}{A data frame produced by \code{combine_reps()} corresponding to experiment 2.}

\item{exp1_id}{Prefix that will annotate experiment 1 columns.}

\item{exp2_id}{Prefix that will annotate experiment 2 columns.}

\item{outfile_prefix}{Optional. File prefix for output \code{.csv} files. Default = "exps_combined".}
}
\value{
A data frame and \code{.csv} with PSM information across biological replicates combined into one table.
}
\description{
Combine different proteomics experiments, i.e., an HA-tag pulldown and a TurboID-biotin pulldown. Compares joint z-scores and FDR across each experiment. See README for file formats and example usage on the \href{https://github.com/rachaelcox/enrichr}{enrichr GitHub Repository}.
}
