% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_fxn}
\alias{compute_fxn}
\title{Compute fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs.}
\usage{
compute_fxn(pd_impute_df, one_sided = FALSE)
}
\arguments{
\item{pd_impute_df}{A data frame generated by \code{impute_fxn()}.}

\item{one_sided}{Specify one-sided or two-sided statistical test; default = \code{FALSE} (two-sided).}
}
\value{
A data frame with fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs.
}
\description{
Compute fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs.
}
