% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{enrich_fxn}
\alias{enrich_fxn}
\title{Annotate and compute fold-changes and Z-scores for differential proteomics results.}
\usage{
enrich_fxn(
  exp_id,
  meta_file,
  psm_file,
  pd_file,
  annot_file,
  outfile_name = "output.csv"
)
}
\arguments{
\item{exp_id}{An identifier for a given experiment that matches test and control files together (see meta file)}

\item{meta_file}{A tab-separate file with column headers \code{exp_name} (populated with values that correspond to exp_id variable), \code{type} (exp or ctrl), and \code{file_id} (determined by Proteome Discoverer, see README on Github)}

\item{psm_file}{PSM.txt tab-separated file output by Proteome Discoverer 2.3}

\item{pd_file}{Proteins.txt tab-separated file output by Proteome Discoverer 2.3}

\item{annot_file}{Optional tab-separated file containing a table with annotations for the queried proteome; must match accession format in FASTA given to Proteome Discoverer as a look up database}

\item{outfile_name}{File name for output .csv; default = "output.csv"}
}
\value{
A .csv with all computed information combined into one table.
}
\description{
Format PSM.txt and Proteins.txt files output by Proteome Discoverer 2.3, formatting PSM counts and AUC XIC values into one table, summing technical replicates, and computing fold-change and Z-scores for each detected protein in test and control runs. Optionally annotates results given a tab-separated annotation file. See README for file formats and example usage on the \href{https://github.com/rachaelcox/enrichr}{enrichr GitHub Repository}.
}
