% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_pd_file}
\alias{format_pd_file}
\title{Format a Proteins.txt file output by Proteome Discoverer 2.3, summing technical replicates and demarcating test and control PSMs}
\usage{
format_pd_file(exp_id, meta_file, pd_file, psm_df)
}
\arguments{
\item{exp_id}{An identifier for a given experiment that matches both test and control files}

\item{meta_file}{A tab-separate file with column headers \code{exp_name} (populated with values that correspond to exp_id variable, \code{type} (exp or ctrl) \code{file_id} (determined by Proteome Discoverer)}

\item{pd_file}{Proteins.txt tab-separated file output by Proteome Discoverer 2.3}

\item{psm_df}{A dataframe output from \code{format_psm_file()}}
}
\value{
A dataframe with AUC XIC values, PSM counts and additional PSM info summarized for each accession in test and control cases
}
\description{
Format a Proteins.txt file output by Proteome Discoverer 2.3, summing technical replicates and demarcating test and control PSMs
}
