% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{impute_fxn}
\alias{impute_fxn}
\title{Impute minimum values for AUC XICs for each accession in both test and control runs.}
\usage{
impute_fxn(pd_df)
}
\arguments{
\item{pd_df}{A dataframe generated by \code{format_pd_file()}.}
}
\value{
A dataframe with imputed values for AUC XICs.
}
\description{
Impute minimum values for AUC XICs for each accession in both test and control runs.
}
