% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{psmplot}
\alias{psmplot}
\title{Visualize control PSMs versus test PSMs on a log-log plot.}
\usage{
psmplot(
  data,
  outfile_prefix,
  threshold,
  num_labs,
  label_file,
  ylab,
  xlab,
  point_color
)
}
\arguments{
\item{data}{A data frame produced by \code{enrich()}.}

\item{outfile_prefix}{Optional. File prefix for output \code{.png} and \code{.pdf} files. Default = "output."}

\item{threshold}{Optional. FDR threshold to color points by. One of c(90, 95, 99). Default = 5\% FDR (fdr_bh <= 0.05).}

\item{num_labs}{Optional. Number of annotation labels to display. Requires input to have an annotation column named \code{gene_names_primary}. Default = 0.}

\item{label_file}{Optional. Comma-separated file containing names of genes to be labeled on the plot; requires input to have an annotation column named \code{gene_names_primary} and labels must match format in the column.}

\item{ylab}{Text for y-axis label; will correspond to test PSMs.}

\item{xlab}{Text for x-axis label; will correspond to control PSMs.}
}
\value{
PSM log-log plots in \code{.png} and \code{.pdf} format.
}
\description{
Graphs control PSMs versus test PSMs on a log-log plot. A pseudocount of 1 is adding to all PSMs for log-log visualization. Colored by Z-test threshold. See README for file formats and example usage on the \href{https://github.com/rachaelcox/diffprot}{diffprot GitHub Repository}.
}
