% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{zplot}
\alias{zplot}
\title{Plot Z-scores across biological replicates.}
\usage{
zplot(
  data,
  outfile_prefix,
  ylab,
  xlab,
  ycol,
  xcol,
  threshold,
  label_file,
  num_labels
)
}
\arguments{
\item{data}{A data frame produced by \code{combine_reps()}.}

\item{outfile_prefix}{Optional. File prefix for output \code{.png} and \code{.pdf} files. Default = "output."}

\item{ylab}{Text for y-axis label; should correspond to replicate #1 Z-scores.}

\item{xlab}{Text for x-axis label; should correspond to replicate #2 Z-scores.}

\item{ycol}{Name of column containing y-axis data; should correspond to replicate #1 Z-scores.}

\item{xcol}{Name of column containing x-axis data; should correspond to replicate #1 Z-scores.}

\item{threshold}{Optional. FDR threshold to color points by. One of c(90, 95, 99). Default = 5\% FDR (fdr_bh <= 0.05).}

\item{label_file}{Optional. Comma-separated file containing names of genes to be labeled on the plot; requires input to have an annotation column named \code{gene_names_primary} and labels must match format in the column.}

\item{num_labs}{Optional. Number of annotation labels to display. Requires input to have an annotation column named \code{gene_names_primary}. Default = 0.}
}
\value{
Plot depicting replicate 1 z-score vs replicate 2 z-scores in \code{.png} and \code{.pdf} format. Colored by FDR.
}
\description{
Colored by a multiple hypothesis corrected false discovery rate (FDR) computed on a weighted Z-score. See README for file formats and example usage on the \href{https://github.com/rachaelcox/enrichr}{enrichr GitHub Repository}.
}
