% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_reps}
\alias{combine_reps}
\title{Combine replicate differential proteomics experiments.}
\usage{
combine_reps(rep1, rep2, rep3, outfile_prefix)
}
\arguments{
\item{rep1}{A data frame produced by \code{enrich_fxn()} corresponding to replicate 1.}

\item{rep2}{A data frame produced by \code{enrich_fxn()} corresponding to replicate 2.}

\item{rep3}{Optional. A data frame produced by \code{enrich_fxn()} corresponding to replicate 3.}

\item{outfile_prefix}{Optional. File prefix for output \code{.csv} files. Default = "output_combined."}
}
\value{
A \code{.csv} with PSM information across biological replicates combined into one table.
}
\description{
Combine replicate proteomics experiments. Calculates mean PSMs across test and control runs. Calculates weighted z-scores using Stouffer's Z-score method. See README for file formats and example usage on the \href{https://github.com/rachaelcox/enrichr}{enrichr GitHub Repository}.
}
