% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_fxn}
\alias{compute_fxn}
\title{Compute fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs}
\usage{
compute_fxn(pd_impute_df)
}
\arguments{
\item{pd_impute_df}{A dataframe generated by \code{impute_fxn()}}
}
\value{
A dataframe with fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs
}
\description{
Compute fold-change values for XICs and PSMs & Z-scores for PSMs for each accession in both test and control runs
}
